refactor(server-queries): make sure file names match style #29
