+++ .changeset/rotten-hotels-mix.md
@@ -1,0 +1,5 @@
+---
+"server-queries": patch
+---
+
+Ensure consistent file names
