+++ packages/server-queries/src/entry/client.ts
@@ -1,10 +1,10 @@
 export { ServerQueryConfigProvider } from "../context/ServerQueryConfigProvider.client";
 export { useServerQueryConfig } from "../context/ServerQueryConfigProvider.hooks";
 
-export { useServerQuery } from "../hooks/useServerQuery";
-export { useServerInfiniteQuery } from "../hooks/useServerInfiniteQuery";
-export { useServerMutation } from "../hooks/useServerMutation";
-export { useServerAction } from "../hooks/useServerAction";
+export { useServerQuery } from "../hooks/use-server-query";
+export { useServerInfiniteQuery } from "../hooks/use-server-infinite-query";
+export { useServerMutation } from "../hooks/use-server-mutation";
+export { useServerAction } from "../hooks/use-server-action";
 export { createCaller } from "../lib/caller";
 
 export type {
