+++ packages/server-queries/src/hooks/mutation/use-throw-on-error-fn.ts
