+++ packages/server-queries/src/hooks/use-server-query.ts
